Pushing Jersh's changes
Added items, strength and max health.

Each item has a vector of AttributeChange s. each of which has an enum
that's either STRENGTH, HP, OR RESRATE and an int value. The value is
how much it increases or decreases that attribute. Later in the week we
can add intelligence or agility or something if just HP and Str is too
simple.

Each character can have 2 items. Use receiveItem() to give them the
goods since that's the only thing that will remove the old item's stats
(if they have an old item) and add the new item's stats.

Also added max health and strength and the drawing of max health (in the
right order, awww yeah(it was originally 100/90 when they took 10 damage
but now it should be good))